+++ packages/cli/src/commands/apply.ts
@@ -1,10 +1,11 @@
 import { ScheduleSchema } from "@enschedule/types";
+import { NetworkError } from "@enschedule/worker-api";
 import { Command } from "commander";
 import { glob } from "glob";
 import yaml from "js-yaml";
 import fs from "node:fs";
 import { z } from "zod";
-import { getAuthHeader, getWorker } from "../get-worker";
+import { ConfigError, getAuthHeader, getWorker } from "../get-worker";
 
 const isDirectory = async (p: string): Promise<boolean> => {
   const stat = await fs.promises.stat(p);
@@ -25,16 +26,34 @@ export const ScheduleYamlSchema = z.object({
 export const apply = async (
   config: z.infer<typeof ScheduleYamlSchema>
 ): Promise<void> => {
-  const authHeader = await getAuthHeader();
-  const worker = await getWorker();
-  const { status } = await worker.scheduleJob(
-    authHeader,
-    config.spec.handlerId,
-    config.spec.handlerVersion,
-    config.spec.data,
-    { ...config.spec.options, eventId: config.metadata.name }
-  );
-  console.log("Schedule", config.metadata.name, status);
+  try {
+    const authHeader = await getAuthHeader();
+    const worker = await getWorker();
+    try {
+      const { status } = await worker.scheduleJob(
+        authHeader,
+        config.spec.handlerId,
+        config.spec.handlerVersion,
+        config.spec.data,
+        { ...config.spec.options, eventId: config.metadata.name }
+      );
+      console.log("Schedule", config.metadata.name, status);
+    } catch (err) {
+      if (err instanceof NetworkError) {
+        console.error("Network error:", err.cliMessage);
+        process.exit(1);
+      } else {
+        throw err;
+      }
+    }
+  } catch (err) {
+    if (err instanceof ConfigError) {
+      console.error(
+        "You have an error in your config. Please check your ~/.enschedule/config.json file or the ENSCHEDULE_API_KEY, ENSCHEDULE_API_ENDPOINT, and ENSCHEDULE_API_VERSION environment variables."
+      );
+      process.exit(1);
+    }
+  }
 };
 
 applyCommand
@@ -70,6 +89,7 @@ applyCommand
         );
       });
 
+    let hasApplied = false;
     for (const absolutePath of resolvedPaths) {
       const fileContent = fs.readFileSync(absolutePath, "utf8");
       let data;
@@ -88,10 +108,16 @@ applyCommand
         for (const config of configs) {
           // eslint-disable-next-line no-await-in-loop
           await apply(config);
+          hasApplied = true;
         }
       } else {
         // eslint-disable-next-line no-await-in-loop
         await apply(ScheduleYamlSchema.parse(data));
+        hasApplied = true;
       }
+    }
+    if (!hasApplied) {
+      console.error("No schedules found in the provided file or folder");
+      process.exit(1)
     }
   });