+++ packages/cli/src/index.ts
@@ -4,9 +4,11 @@ import { applyCommand } from "./commands/apply";
 import { getCommand } from "./commands/get";
 import { deleteCommand } from "./commands/delete";
 import pkg from "../package.json";
+import { resetCommand } from "./commands/reset";
 
 program.addCommand(applyCommand);
 program.addCommand(getCommand);
 program.addCommand(deleteCommand);
+program.addCommand(resetCommand);
 
 program.version(pkg.version).parse(process.argv);