+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @enschedule/dashboard
 
+## 1.1.0
+
+### Minor Changes
+
+- Improve error and auth handling related to the CLI
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/worker-api@1.1.0
+  - @enschedule/worker@1.1.0
+  - @enschedule/types@1.1.0
+
 ## 1.0.0
 
 ### Major Changes