+++ packages/hub/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @enschedule/hub
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/worker-api@1.1.0
+  - @enschedule/pg-driver@1.1.0
+  - @enschedule/worker@1.1.0
+  - @enschedule/dashboard@1.1.0
+  - @enschedule/types@1.1.0
+
 ## 1.0.0
 
 ### Major Changes