+++ packages/pg-driver/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @enschedule/pg-driver
 
+## 1.1.0
+
+### Minor Changes
+
+- Improve error and auth handling related to the CLI
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/types@1.1.0
+
 ## 1.0.0
 
 ### Major Changes