+++ packages/worker-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enschedule/worker-api
 
+## 1.1.0
+
+### Minor Changes
+
+- Improve error and auth handling related to the CLI
+
 ## 1.0.0
 
 ### Major Changes