+++ tests/hub-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enschedule/hub-test
 
+## 0.0.3
+
+### Patch Changes
+
+- @enschedule/hub@1.0.1
+
 ## 0.0.2
 
 ### Patch Changes